refactor(clawd): extract EnvironmentLoader and DaemonBuilder from RunCommand
The whole service graph hung off the CLI command type, and doctor
re-implemented the environment preamble (config load, secret resolution,
store open) that run's *OrExit helpers also performed — two copies that
would drift. EnvironmentLoader now holds the single implementation of each
bootstrap step (plus the shared database/workspace path derivations); run
consumes it through thin exit-code-mapping wrappers and doctor calls the
steps individually for its per-row diagnostics. The graph assembly moves
onto DaemonBuilder, organized by subsystem (Agent, Approvals, Intake,
Schedule, Boot), keeping the single composition root and the make* naming.
RunCommand keeps argument parsing and process lifecycle only.
